+++ RELEASENOTES.md
@@ -3,11 +3,9 @@ These are the release notes for the [mediawiki-api](README.md).
 ## Version 0.3 (under development)
 
 * Removes NewEditInfo and NewRevision
-* Introduces ServiceFactory
+* Introduces MediawikiFactory
 * PageRepo, RevisionSaver and UserRepo moved to the Service Namespace
-* Correctly handle non existent users in UserRepo
 * Moved basic api functionality to a separate base lib (mediawiki-api-base)
-* Mediawiki\Api\ServiceFactory moved to Mediawiki\Api\Service\ServiceFactory
 * Introduces PageDeleter service
 * Introduces PageListRepo service
 * Introduces PageProtector service
@@ -19,6 +17,7 @@ These are the release notes for the [mediawiki-api](README.md).
 * Introduces UserRightsChanger service
 * Introduces PageRestorer service
 * Introduces RevisionRestorer service
+* Correctly handle non existent users in UserRepo
 
 
 ## Version 0.2 (2014-02-23)